+++ build.gradle
@@ -3,6 +3,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.0.10.RELEASE'
     id 'org.asciidoctor.convert' version '1.5.8'
     id 'java'
+    id 'org.jetbrains.kotlin.jvm' version '1.4.30'
 }
 
 group = 'com.hongikbros'
@@ -55,6 +56,7 @@ dependencies {
     testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
 
     testImplementation 'org.mockito:mockito-inline:2.13.0'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
 }
 
 ext {
@@ -106,5 +108,15 @@ processTestResources {
         from('backend-secret')
         include('resources-test/**')
         into('src/test')
+    }
+}
+compileKotlin {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
+compileTestKotlin {
+    kotlinOptions {
+        jvmTarget = "1.8"
     }
 }
