+++ src/main/java/com/hongikbros/jobmanager/common/config/JpaAuditingConfig.java
@@ -12,7 +12,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-import com.hongikbros.jobmanager.member.application.oauth.CustomOAuth2UserService;
+import com.hongikbros.jobmanager.security.oauth2.CustomOAuth2UserService;
 
 @Configuration
 @EnableJpaAuditing(auditorAwareRef = "auditorProvider")
