+++ src/main/java/com/hongikbros/jobmanager/notice/company/CompanyRepository.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.notice.domain.company;
+package com.hongikbros.jobmanager.notice.company;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
