+++ src/main/java/com/hongikbros/jobmanager/notice/notice/domain/NoticeDescription.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.notice.domain.notice;
+package com.hongikbros.jobmanager.notice.notice.domain;
 
 import java.util.Objects;
