+++ src/main/java/com/hongikbros/jobmanager/notice/notice/domain/NoticeRepository.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.notice.domain.notice;
+package com.hongikbros.jobmanager.notice.notice.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
