+++ src/main/java/com/hongikbros/jobmanager/security/oauth2/CustomOauth2SuccessHandler.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.common.domain.security;
+package com.hongikbros.jobmanager.security.oauth2;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -16,7 +16,6 @@ import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hongikbros.jobmanager.member.application.member.MemberResponse;
-import com.hongikbros.jobmanager.member.application.oauth.CustomOAuth2UserService;
 import com.hongikbros.jobmanager.member.ui.SessionMember;
 
 @Component
