+++ src/main/java/com/hongikbros/jobmanager/skill/domain/SkillRepository.java
@@ -1,10 +1,9 @@
 package com.hongikbros.jobmanager.skill.domain;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SkillRepository extends JpaRepository<Skill, Long> {
-    Optional<List<Skill>> findByIdIn(List<Long> ids);
+    List<Skill> findByIdIn(List<Long> ids);
 }
