+++ src/test/java/com/hongikbros/jobmanager/security/oauth2/CustomOauth2SuccessHandlerTest.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.common.domain.security;
+package com.hongikbros.jobmanager.security.oauth2;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
@@ -23,7 +23,6 @@ import org.springframework.security.core.Authentication;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hongikbros.jobmanager.common.fixture.member.MemberFixture;
 import com.hongikbros.jobmanager.member.ui.SessionMember;
-import com.hongikbros.jobmanager.member.application.oauth.CustomOAuth2UserService;
 
 @ExtendWith(MockitoExtension.class)
 class CustomOauth2SuccessHandlerTest {
